+++ CHANGELOG.md
@@ -53,6 +53,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Preservation of the original token file positions
 - Enhance `MixedNamesRule` with symbol table information.
 - Exclude class constructors in `ConstructorCreateRule`.
+- Exclude untyped expression casts in `CastAndFreeRule`.
 - Handle `else` exception handler in `SwallowedExceptionsRule`.
 - Handle global variables separately in `VariableNameRule`.
 - Allow empty case branches in `EmptyBeginStatementRule` if they have comments.